MK-1359 Add harmonization-study
